requirements.txt
beautifulsoup4==4.12.3
langchain==0.3.8
langchain_community==0.3.8
langchain_huggingface==0.1.2
ollama==0.4.1
pdfplumber==0.11.4
Pillow==11.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.26.1
sentence_transformers==3.3.1
torch==2.5.1+cu118
tqdm==4.67.0
tf-keras==2.18.0
faiss-cpu==1.9.0.post1
llama-cpp-python==0.1.77
unstructured==0.8.0
pymupdf==1.22.5
python-docx==0.8.11
Flask==3.1.0
